refactoring of scripts
