+++ databases/catdat/scripts/proof-length.ts
@@ -1,5 +1,5 @@
 import { type StructureType } from './config'
-import { get_client } from './utils/helpers'
+import { get_client } from './utils/db'
 
 const db = get_client()
